requirements.txt
ijson
grequests
joblib
dplython
appdirs==1.4.3
astroid==1.5.2
backports.functools-lru-cache==1.4
BeautifulSoup==3.2.1
configparser==3.5.0
cssselect==1.0.1
dj-database-url==0.4.1
Django==1.11.1
django-debug-toolbar==1.8
enum34==1.1.6
goose-extractor==1.0.25
gunicorn==19.6.0
isort==4.2.5
jieba==0.38
lazy-object-proxy==1.3.1
lxml==3.7.3
mccabe==0.6.1
nltk==3.2.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.6.2
pylint==1.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.3
whitenoise==3.2
wrapt==1.10.10
gensim==2.1.0
nltk==3.2.2
numpy==1.12.1
scikit-learn==0.18.1
sklearn==0.0
mysqlclient==1.3.12
